fixed order reducer
